Safari-611.1.10/LayoutTests/inspector/css/manager-preferredInspectorStyleSheetForFrame-expected.txt

Test CSSManager.preferredInspectorStyleSheetForFrame.


PASS: Page should have a subframe.

== Running test suite: CSSManager.preferredInspectorStyleSheetForFrame
-- Running test case: CheckNoStyleSheets
PASS: Should be no stylesheets.

-- Running test case: CreateMainFrameInspectorStyleSheet
PASS: Added StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
PASS: Added StyleSheet frame should be the main frame.
PASS: Should be one stylesheet.
PASS: StyleSheet origin is inspector.
PASS: StyleSheet.isInspectorStyleSheet() should be true.

-- Running test case: AnotherMainFrameRequestDoesNothing
PASS: Should still be one stylesheet.
PASS: StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.

-- Running test case: CreateSubFrameInspectorStyleSheet
PASS: Added StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
PASS: Added StyleSheet frame should be a child frame.
PASS: Should be two stylesheets.
PASS: StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.

-- Running test case: AnotherSubFrameRequestDoesNothing
PASS: Should be two stylesheets.
PASS: StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
